Split `storage` module
